EUI-5068-case-flags
