+++ api/hearings/data/serviceHearingValues.mock.data.ts
@@ -120,9 +120,9 @@ export const SERVICE_HEARING_VALUES: ServiceHearingValuesModel = {
       },
       {
         partyName: 'Jane Smith',
-        flagParentId: 'CF0001',
-        flagId: 'CF0007',
-        flagDescription: 'Urgent flag',
+        flagParentId: 'PF0001',
+        flagId: 'PF0002',
+        flagDescription: 'Vulnerable user',
         flagStatus: 'ACTIVE',
       },
       {
@@ -134,30 +134,30 @@ export const SERVICE_HEARING_VALUES: ServiceHearingValuesModel = {
       },
       {
         partyName: 'DWP',
-        flagParentId: 'CF0001',
-        flagId: 'CF0002',
-        flagDescription: 'Complex Case',
-        flagStatus: 'ACTIVE',
-      },
-      {
-        partyName: 'DWP',
-        flagParentId: 'PF0001',
-        flagId: 'PF0002',
-        flagDescription: 'Vulnerable user',
-        flagStatus: 'ACTIVE',
-      },
-      {
-        partyName: 'DWP',
         flagParentId: 'PF0001',
         flagId: 'PF0011',
         flagDescription: 'Banning order',
         flagStatus: 'ACTIVE',
       },
       {
-        partyName: 'DWP',
+        partyName: 'Jane Smith v DWP',
+        flagParentId: 'CF0001',
+        flagId: 'CF0002',
+        flagDescription: 'Complex Case',
+        flagStatus: 'ACTIVE',
+      },
+      {
+        partyName: 'Jane Smith v DWP',
         flagParentId: 'CF0001',
         flagId: 'CF0006',
         flagDescription: 'Potential fraud',
+        flagStatus: 'ACTIVE',
+      },
+      {
+        partyName: 'Jane Smith v DWP',
+        flagParentId: 'CF0001',
+        flagId: 'CF0007',
+        flagDescription: 'Urgent flag',
         flagStatus: 'ACTIVE',
       },
     ],
